NavigationCompose.kt added, separated file
